requirements.txt
# PBR should always appear first
pbr>=1.6 # Apache-2.0

# iSwfit requirements
Babel>=1.3 # BSD
Django<1.9,>=1.8 # BSD
django-openstack-auth>=2.0.0 # Apache-2.0

httplib2>=0.7.5 # MIT
iso8601>=0.1.9 # MIT

python-keystoneclient!=1.8.0,!=2.1.0,>=1.6.0 # Apache-2.0
python-swiftclient>=2.2.0 # Apache-2.0
